liveness_defense/requirements.txt
opencv-python
mediapipe
numpy
scipy
matplotlib
python-dotenv
certifi